refactoring test: rename gym_api_test.py to f110_env_test.py, move test configure() in f110_env_test.py, refactoring utility test
